Get acceptance specs running
