+++ config/locales/metric.yml
@@ -1,8 +1,4 @@
-# Sample localization file for English. Add more files in this directory for other locales.
-# See https://github.com/svenfuchs/rails-i18n/tree/master/rails%2Flocale for starting points.
-
 metric:
-  
   models:
     calculator:                 "Calculator"
     crop:                       "Crop"
@@ -60,7 +56,7 @@ metric:
   layouts:
     application:
       title:          "Fertiliser And Seed Decision Aid Calculator"
-      
+
   calculator:
     index:
       title:          "Seed Damage Calculator"
