+++ spec/support/blueprints.rb
@@ -2,36 +2,36 @@ require 'machinist/active_record'
 require 'faker'
 
 Crop.blueprint do
-  name    { "#{Faker::Lorem.words.first}-#{sn}" }
+  name    { "Crop-#{sn}" }
 end
 
 Fertiliser.blueprint do
-  metric_name     { "#{Faker::Lorem.words.first}-#{sn}" }
-  imperial_name   { "#{Faker::Lorem.words.first}-#{sn}" }
-  N               { rand().round(2) }
-  P               { rand().round(2) }
-  K               { rand().round(2) }
-  S               { rand().round(2) }
-  Mg              { rand().round(2) }
-  liquid_weight   { rand().round(6) }
+  metric_name     { "met-fert-#{sn}" }
+  imperial_name   { "imp-fert-#{sn}" }
+  N               { 0.33 }
+  P               { 0.33 }
+  K               { 0.33 }
+  S               { 0.33 }
+  Mg              { 0.33 }
+  liquid_weight   { 0.334329 }
 end
 
 SoilTexture.blueprint do
-  name            { "#{Faker::Lorem.words.first}-#{sn}" }
+  name            { "fine" }
 end
 
 SoilMoisture.blueprint do
-  name            { "#{Faker::Lorem.words.first}-#{sn}" }
+  name            { "moist" }
 end
 
 SoilCoefficient.blueprint do
-  value           { rand().round(2) }
+  value           { 1.5 }
   soil_texture
   soil_moisture
 end
 
 RegressionCoefficient.blueprint do
-  value           { Random.new.rand(-5..5) + rand() }
+  value           { -0.131 }
   crop
   fertiliser
 end
